+++ exonum-java-binding-common/src/test/java/com/exonum/binding/common/proofs/map/MapEntryTest.java
@@ -16,6 +16,8 @@
 
 package com.exonum.binding.common.proofs.map;
 
+import static com.google.common.testing.NullPointerTester.Visibility;
+
 import com.exonum.binding.common.collect.MapEntry;
 import com.google.common.testing.NullPointerTester;
 import com.google.protobuf.ByteString;
@@ -23,12 +25,12 @@ import nl.jqno.equalsverifier.EqualsVerifier;
 import nl.jqno.equalsverifier.Warning;
 import org.junit.jupiter.api.Test;
 
-class ByteStringMapEntryTest {
+class MapEntryTest {
 
   @Test
-  void constructorRejectsNulls() {
+  void staticFactoryMethodsRejectsNulls() {
     new NullPointerTester()
-        .testAllPublicConstructors(MapEntry.class);
+        .testStaticMethods(MapEntry.class, Visibility.PUBLIC);
   }
 
   @Test
@@ -37,7 +39,7 @@ class ByteStringMapEntryTest {
         .usingGetClass()
         .suppress(Warning.INHERITED_DIRECTLY_FROM_OBJECT)
         .withPrefabValues(
-    ByteString.class, ByteString.copyFromUtf8("a"), ByteString.copyFromUtf8("b"))
+            ByteString.class, ByteString.copyFromUtf8("a"), ByteString.copyFromUtf8("b"))
         .verify();
   }